Get the basic idea working
The code ain't pretty ... but it works.
Need to verify that the setup of the token still works at some point.
The getRows file is most of the Google Quickstart example, with some
added helper functions and broken apart a bit.
It's pretty ugly right now but not really the focus.